first commit of the e2e framework enhancements
