+++ frontend/fingertips-frontend/playwright/testHelpers/genericTestUtils.ts
@@ -1,6 +1,6 @@
-export * from './testHelpers/testEnums';
-export * from './testHelpers/indicatorUtils';
-export * from './page-objects/components/scenarioMapper';
+export * from './testEnums';
+export * from './indicatorUtils';
+export * from '../page-objects/components/scenarioMapper';
 
 export function sortAlphabetically(array: (string | null)[]) {
   array.sort((a, b) => a!.localeCompare(b!));
