+++ frontend/fingertips-frontend/playwright/tests/fully_integrated_e2e_tests/ai_search_tests/area_search.spec.ts
@@ -1,6 +1,6 @@
 import { AreaDocument } from '@/lib/search/searchTypes';
 import { test } from '../../../page-objects/pageFactory';
-import { TestTag } from '@/playwright/testHelpers';
+import { TestTag } from '@/playwright/testHelpers/genericTestUtils';
 
 test.describe(
   'Azure AI Search: Area search suggestions',
