+++ frontend/fingertips-frontend/playwright/tests/fully_integrated_e2e_tests/ai_search_tests/indicator_search.spec.ts
@@ -1,4 +1,4 @@
-import { SearchMode, TestTag } from '@/playwright/testHelpers';
+import { SearchMode, TestTag } from '@/playwright/testHelpers/genericTestUtils';
 import { test } from '../../../page-objects/pageFactory';
 
 test.describe(
